refactor(app): extract prompt history navigation
Move two persisted history stores, historyComments, applyHistoryComments,
applyHistoryPrompt, addToHistory, and navigateHistory into
prompt-input/history-navigation.ts.

History is a self-contained subsystem: persisted entries (normal + shell),
apply logic with comment synchronization, and navigation. setCursorPosition
imports directly from editor-dom (no detour through editor-imperatives).

Adopt PromptStore type from store-types in the main createStore call.
